rename: api 폴더명을 apis로 변경
- 다른 폴더들과의 통일성을 위함
Fixes: #84
