+++ src/pages/auth/SignUpPage.tsx
@@ -7,7 +7,7 @@ import { Browser } from 'components/common/Browser';
 import { Input } from 'components/common/Input';
 import { Button } from 'components/common/Button';
 import WindowsImg from 'assets/windows_img.png';
-import { auth } from 'api/auth';
+import { auth } from 'apis/auth';
 import { useValidInput } from 'hooks/useValidInput';
 import { CustomError } from 'global/types';
