+++ src/pages/board/ListPage.tsx
@@ -5,8 +5,8 @@ import { Browser } from 'components/common/Browser';
 import { Layout } from 'components/layouts/Layout';
 import { CustomError, Post } from 'global/types';
 import { Button } from 'components/common/Button';
-import { auth } from 'api/auth';
-import { board } from 'api/board';
+import { auth } from 'apis/auth';
+import { board } from 'apis/board';
 import { PostItem } from 'components/board/PostItem';
 import { NoPost } from 'components/common/NoPost';
 import { useIntersectionObserver } from 'hooks/useIntersectionObserver';
